requirements.txt
bcrypt==4.3.0
blinker==1.9.0
click==8.2.0
dnspython==2.7.0
dotenv==0.9.9
email_validator==2.2.0
Flask==3.1.1
flask-cors==5.0.1
Flask-PyMongo==3.0.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.0
MarkupSafe==3.0.2
numpy==2.2.5
pymongo==4.13.0
python-dotenv==1.1.0
scikit-learn==1.6.1
scipy==1.15.3
threadpoolctl==3.6.0
Werkzeug==3.1.3
